+++ Controller/RoleController.php
@@ -75,25 +75,40 @@
 		 * 选择要执行哪个动作
 		 */
 		public function selectAction(){
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="addRole"){
-				return $this->addRole();
+			//判断有没有请求动作，因为有php页面直接调用
+			if(isset($_REQUEST['action'])){
+				//用户需要登录系统，并且有权限才能执行相应的action
+				if($this->roleManager->isUserLogon()){
+					//可以执行的action
+					if($this->roleManager->hasAuthority(RoleManage)){
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="addRole"){
+							return $this->addRole();
+						}
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadRole"){
+							return $this->loadRole();
+						}
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadAuthorityInfo"){
+							return $this->loadAuthorityInfo();
+						}
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadRoleInfoByRoleId"){
+							return $this->loadRoleInfoByRoleId();
+						}
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="changeRoleInfo"){
+							return $this->changeRoleInfo();
+						}
+						if(isset($_REQUEST['action']) && $_REQUEST['action']=="deleteRole"){
+							return $this->deleteRole();
+						}
+					}
+					//否则返回无权限信息
+					else{
+						return NoAuthority;
+					}
+				}
+				else{
+					return NotLogon;
+				}
 			}
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadRole"){
-				return $this->loadRole();
-			}
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadAuthorityInfo"){
-				return $this->loadAuthorityInfo();
-			}
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="loadRoleInfoByRoleId"){
-				return $this->loadRoleInfoByRoleId();
-			}
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="changeRoleInfo"){
-				return $this->changeRoleInfo();
-			}
-			if(isset($_REQUEST['action']) && $_REQUEST['action']=="deleteRole"){
-				return $this->deleteRole();
-			}
-			return "没有发送合适的请求";
 		}
 	}
